requirements-dev.txt
black==25.9.0
click==8.3.0
flake8==7.3.0
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytokens==0.1.10
PyYAML==6.0.3
typing_extensions==4.15.0
yamllint==1.37.1
django-stubs==5.2.5
djangorestframework-stubs==3.16.4
pytest-django==4.11.1
setuptools==74.0.0
python-semantic-release==10.4.1